refactor: 사용 구분에 따른 패키지명 재 구성
