+++ src/main/java/racing/controller/RacingController.java
@@ -1,11 +1,11 @@
 package racing.controller;
 
 import java.util.List;
-import racing.dto.RacingInput;
-import racing.dto.RacingOutput;
+import racing.domain.dto.RacingInput;
+import racing.domain.dto.RacingOutput;
 import racing.service.RacingService;
-import racing.strategy.NumberCompareMoveStrategy;
-import racing.strategy.RandomNumberGenerator;
+import racing.domain.strategy.NumberCompareMoveStrategy;
+import racing.domain.strategy.RandomNumberGenerator;
 import racing.view.RacingInputView;
 import racing.view.RacingOutputView;
 
@@ -15,7 +15,8 @@ public class RacingController {
     RacingInputView racingInputView = new RacingInputView();
     RacingInput racingInput = racingInputView.printAndGetRacingInput();
 
-    RacingService racingService = new RacingService(new NumberCompareMoveStrategy(new RandomNumberGenerator()));
+    RacingService racingService = new RacingService(
+        new NumberCompareMoveStrategy(new RandomNumberGenerator()));
     List<RacingOutput> racingOutputs = racingService.race(racingInput);
 
     RacingOutputView racingOutputView = new RacingOutputView(racingOutputs);
