+++ src/main/java/racing/domain/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racing.strategy;
+package racing.domain.strategy;
 
 public interface MoveStrategy {
