+++ src/main/java/racing/domain/strategy/NumberCompareMoveStrategy.java
@@ -1,4 +1,4 @@
-package racing.strategy;
+package racing.domain.strategy;
 
 public class NumberCompareMoveStrategy implements MoveStrategy {
