+++ src/test/java/racing/domain/strategy/FixedNumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package racing.strategy;
+package racing.domain.strategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
